requirements.txt
# Core libraries
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
httpx==0.26.0

# Microsoft Graph API
msal==1.26.0
pyngrok==7.0.5

# Database
pymongo==4.6.1

# AI/ML
openai==1.12.0
torch==2.0.1
transformers==4.37.2
pandas==2.1.0

# Email processing
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
retry==0.9.2


# Core Utilities
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
httpx==0.26.0
python-dateutil==2.8.2
tqdm==4.66.1
retry==0.9.2

# Environment & System
typing-extensions==4.10.0
email-validator==2.1.0

# Microsoft Graph API & Authentication
msal==1.26.0
pyngrok==7.0.5

# Database Access
pymongo==4.6.1
bson==0.5.10

# AI/ML Libraries
openai==1.12.0
torch==2.0.1
transformers==4.37.2
spacy==3.7.2
sentence-transformers==2.6.1
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.1.0

# Concurrency & Caching
concurrent-log-handler==0.9.24
cachetools==5.3.3  # used internally by msal and useful in caching predictions

# Optional Logging Enhancer
loguru==0.7.2
